+++ client/components/Auth/Login.js
@@ -8,9 +8,9 @@ import {
   TouchableHighlight,
 } from 'react-native'
 import { Container, Form, Item, Input, Content } from 'native-base'
-import { auth } from '../store'
+import { auth } from '../../store'
 import { connect } from 'react-redux'
-import { devEmail } from '../secrets'
+import { devEmail } from '../../secrets'
 
 class Login extends React.Component {
   constructor() {
