+++ client/components/Comment/AddComment.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { StyleSheet, Text, View, Button, Image, KeyboardAvoidingView } from 'react-native'
 import { Form, Textarea } from 'native-base'
 import { connect } from 'react-redux'
-import { postCommentThunk, getCommentsThunk } from '../store'
+import { postCommentThunk, getCommentsThunk } from '../../store'
 
 class AddComment extends React.Component {
   constructor(props) {
