+++ client/components/index.js
@@ -1,26 +1,25 @@
-export { default as ProgressChart } from './ProgressChart'
-export { default as ActivityChart } from './ActivityChart'
-export { default as Login } from './Login'
-export { default as AddActivity } from './AddActivity'
+export { default as ProgressChart } from './DataVisualization/ProgressChart'
+export { default as ActivityChart } from './DataVisualization/ActivityChart'
+export { default as Login } from './Auth/Login'
+export { default as AddActivity } from './Activity/AddActivity'
 export { default as Camera } from './Camera'
-export { default as MapComp } from './MapComp'
-export { default as Map } from './map-wrapper'
-// export { default as Dashboard } from './Dashboard'
-export { default as Activity } from './Activity'
+export { default as MapComp } from './Map/MapComp'
+export { default as Map } from './Map/map-wrapper'
+export { default as Activity } from './Activity/Activity'
 export { default as LoadingScreen } from './LoadingScreen'
-export { default as ActivityCard } from './ActivityCard'
-export { default as Leaderboard } from './Leaderboard'
-export { default as Friends } from './Friends'
-export { default as SocialMedia } from './Social-Media'
-export { default as SearchUsers } from './search-users'
-export { default as AddComment } from './AddComment'
-export { default as AddAd } from './AddAd'
-export { default as CommentCard } from './CommentCard'
-export { default as AdView } from './AdView'
-export { default as Logout } from './Logout'
-export { default as Feed } from './feed'
-export { default as FeedCard } from './feed-card'
-export { default as UserDashboard } from './user-dashboard'
-export { default as FriendDashboard } from './friend-dashboard'
-export { default as Dashboard } from './Dashboard'
-export { default as Marketplace } from './marketplace'
+export { default as ActivityCard } from './Activity/ActivityCard'
+export { default as Leaderboard } from './Leaderboard/Leaderboard'
+export { default as Friends } from './Friends/Friends'
+export { default as SocialMedia } from './Social Media/Social-Media'
+export { default as SearchUsers } from './Search Users/search-users'
+export { default as AddComment } from './Comment/AddComment'
+export { default as AddAd } from './Ad/AddAd'
+export { default as CommentCard } from './Comment/CommentCard'
+export { default as AdView } from './Ad/AdView'
+export { default as Logout } from './Auth/Logout'
+export { default as Feed } from './Feed/feed'
+export { default as FeedCard } from './Feed/feed-card'
+export { default as UserDashboard } from './Dashboard/user-dashboard'
+export { default as FriendDashboard } from './Friends/friend-dashboard'
+export { default as Dashboard } from './Dashboard/Dashboard'
+export { default as Marketplace } from './Marketplace/marketplace'
